fix stupid bug
