+++ src/app/@atasan/kelola-laporan-staf/page.tsx
@@ -1,0 +1,9 @@
+import _Page from "./_page";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+
+export default async function Page() {
+  const session = await getServerSession(authOptions);
+
+  return <_Page session={session} />;
+}
