+++ src/app/@staf/kelola-laporan/buat/page.tsx
@@ -1,0 +1,7 @@
+import { Container } from "@mui/material";
+import SectionBuatLaporan from "./_page";
+import _Page from "./_page";
+
+export default function Page() {
+  return <_Page />;
+}
