--- src/app/@staf/laporan/buat/page.tsx
@@ -1,12 +1,0 @@
-import { Container } from "@mui/material";
-import SectionBuatLaporan from "./_client/SectionBuatLaporan";
-
-export default function Page() {
-  return (
-    <>
-      <Container>
-        <SectionBuatLaporan />
-      </Container>
-    </>
-  );
-}
