+++ src/app/_common/Unauthorized.tsx
@@ -1,0 +1,3 @@
+export default function Unauthorized() {
+  return <div>Unauthorized</div>;
+}
