+++ src/app/api/laporan/route.ts
@@ -3,7 +3,7 @@ import prisma from "@/config/prisma";
 import { Prisma } from "@prisma/client";
 import { getServerSession } from "next-auth";
 import { authOptions } from "../auth/[...nextauth]/route";
-import { TFormBuatLaporan } from "@/app/@staf/laporan/buat/_client/FormBuatLaporan";
+import { TFormBuatLaporan } from "@/app/@staf/kelola-laporan/buat/_client/FormBuatLaporan";
 
 export async function GET(req: NextRequest) {
   const foundAllLaporan = await prisma.laporan.findMany();
