+++ src/app/api/user/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/config/prisma";
 import { User } from "@prisma/client";
 import bcrypt from "bcrypt";
-import { TFormTambahUser } from "@/app/@admin/user/tambah/_client/FormTambahUser";
+import { TFormTambahUser } from "@/app/@admin/kelola-user/tambah/_client/FormTambahUser";
 
 export async function POST(req: NextRequest) {
   const reqData = (await req.json()) as TFormTambahUser;
